+++ ui/src/entities/keys/model/stores/theme.ts
@@ -1,5 +1,7 @@
 import { combine, createEvent, createStore } from 'effector'
 
+import { attachStorage } from '$shared/lib'
+
 import { type KeyboardThemeName, keyboardThemes } from '../../lib/themes'
 
 export const keyboardThemeUpdated = createEvent<KeyboardThemeName>('keyboardThemeUpdated')
@@ -10,3 +12,8 @@ export const keyboardThemeNameStore = createStore<KeyboardThemeName>('defaultDar
 export const keyboardThemeStore = combine(keyboardThemeNameStore, (theme) => keyboardThemes[theme])
 
 keyboardThemeNameStore.on(keyboardThemeUpdated, (_, theme) => theme)
+
+attachStorage({
+  source: keyboardThemeNameStore,
+  key: 'keyboard-theme'
+})