+++ ui/src/shared/lib/index.ts
@@ -1,2 +1,3 @@
 export * from './from-wails-event'
+export * from './local-stored'
 export * from './strings'